requirements-dev.txt
# Development Requirements
# Install with: pip install -r requirements-dev.txt

-r requirements.txt

# Development Tools
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.32.0
jupyter==1.1.1

# Code Quality
black==25.1.0
flake8==7.1.1
isort==6.0.1

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.4.0

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2

# Monitoring and Debugging
django-silk==5.3.0
sentry-sdk==2.21.0

# Static Analysis
mypy==1.14.1
django-stubs==5.1.2
